Add files via upload
This repository has several Java programs that demonstrate different ways to work with arrays. It's a great resource for learning about arrays in Java. That contains programs that reverse the order of elements in an array and methods for adding elements, removing elements, updating elements, and finding index of elements . There are also examples of sorting methods like bubble sort, insertion sort, and selection sort. Additionally, it includes techniques for finding elements like linear search.